dev: Add panel to handle multidimensional image data
The panel name/type is 'ndim'.
The data can be loaded from code or from disk.
imageio is used except for hdf5 files, for which h5py is used.

Dimension names and scales when provided in hdf5 are also loaded and
displayed.

Saving to disk in a numpy, hdf5 or video format is also supported.