dev: Base 'interprete' console background on 'base' color
This is white (#ffffff) for the light windows theme and a dark gray (#2d2d2d) for the dark windows theme. For pyside2 and older pyside6 it remains white even if the dark windows theme is active.

This makes the dark theme at least usable, but it still does not look great.